Remove double type convertations

+++ packages/polygon-packer/src/wasm-nesting.ts
@@ -76,11 +76,11 @@ export default class WasmNesting {
         return this.takeObject(ret) as Float32Array;
     }
 
-    public wasm_packer_get_placement_data(generated_nfp: Float32Array): Uint8Array {
+    public wasm_packer_get_placement_data(generated_nfp: Float32Array): Float32Array {
         const ptr0 = this.passMem(generated_nfp, this.#wasm.__wbindgen_export_1);
         const len0 = this.#vecLen;
         const ret = this.#wasm.wasm_packer_get_placement_data(ptr0, len0);
-        return this.takeObject(ret) as Uint8Array;
+        return this.takeObject(ret) as Float32Array;
     }
 
     public wasm_packer_get_placement_result(placements: Float32Array): Uint8Array {
